\name{isoDu}
\alias{isoDu}
\title{
 Differential isoform expression or usage
}
\description{
 Assess differential isoform expression (DIE) or differential isoform usage (DIU) 
 with respect to a covariate. 
}
\usage{
isoDu(tags, gL, xData, readDepth, outputFileName, pdDistL, g2test, 
  readLen, method=c("bootstrap", "permutation"), nlambda=10, ntau=3, 
  lmax=500, duOnly=FALSE, nResample=10000, pUpper=0.05, pConf=0.99, 
  misAligns=rep(0.001, length(tags)), muMin = 0.01, maxTime=Inf, 
  saveRData="", simulationPhi=NULL, simuNULL=FALSE)
}
\arguments{
  \item{tags}{
  a character vector of the tags of the samples
}
  \item{gL}{
  a list of gene models that are generated from function \code{isoDetector}. The elements 
  of this list is one-to-one correspondence to samples specified in \code{tags}
}
  \item{xData}{
  data of the covariates of interest. \code{xData} could be a vector or a matrix. 
  If it is a vector, its i-th element is the value of the covariate for sample
  tags[i]. 
  
  If xData is a matrix, each column of \code{xData} corresponds to one covariate
  of interest, and the number of rows of \code{xData} should match the length 
  of \code{tags} with its i-th row corresponding to sample tags[i]. 
  We are interested the smallest p-value for all the covariates in xData, and 
  bootstrap or permutation p-value will be calculated for the smallest p-value. 
}
  \item{readDepth}{
  a numeric vector of read depth per sample. It must be one-to-one correspondence to 
  the sample tags in parameter \code{tags}. 
}
  \item{outputFileName}{
  the name of output file.
}
  \item{pdDistL}{
  a list where each element includes the fragment length distribution for a sample. 
  The elements of pdDistL should have one-to-one correspondence with the samples in 
  vector \code{tags}. This list can be generated by function readFragmentSizes.
}
  \item{g2test}{
  the names of the transcript clusters to be tested.
}
  \item{readLen}{
  the length of a read, e.g., 75.
}
  \item{method}{
  resampling method. Must be 'bootstrap' if sample size is smaller than 5.
}  
  \item{nlambda}{
  number of values of tuning parameter lambda. Note that by default we employ a log 
  penalty for coefficient b in form of lambda * log(b + tau).
}
  \item{ntau}{
  number of values of tuning parameter tau. 
}
  \item{lmax}{
  maximum fragment length.
}
  \item{duOnly}{
  if duOnly is TRUE, we test differential isoform usage and otherwise we test differential
  isoform expression.
}
  \item{nResample}{
  number of bootstraps. If nResample <=0, no bootstrapping will be done. 
}
  \item{pUpper}{
  stop the bootstrapping if we have \code{pConf} confidence that the p-value
  is larger than pUpper.
}
  \item{pConf}{
  see the explanation for \code{pConf}.
}
  \item{misAligns}{
  the rate of mis-aligned sequences for samples in group A. This should be 
  a vector of the same length as tags. misAligns times sequence length 
  (in bps) is the effective length of a sequence due to alignment error.
}
  \item{muMin}{
  mu is the expected value of read count for an exon set, which is modelled 
  by a negative binomial distribution. if mu=0, we set mu=muMin. This is used 
  during GLM fitting to improve stability of the algorithm. It should be set to 
  be the same value as is used by function geneModel.
}
  \item{maxTime}{
  maximum amount of CPU time charged for execution by the system 
  (system time) for each transcript cluster. When it reaches the 
  maxTime, the bootstrapping process is stopped.
}
  \item{saveRData}{
  if saveRData != "", the null and alternative model fitting results are written 
  into RData file starting with the character specified by saveRData.
}
  \item{simulationPhi}{
  if !is.null(simulationPhi), carry out isoDU computation using simulated 
  data with over-dispersion parameter specified by simulationPhi.
}  
	\item{simuNULL}{
  if !is.null(simulationPhi), simuNULL specifies to simulate from null or 
  alternative distribution. If simulationPhi is NULL, simuNULL is not used.
}
}
\value{
Results are written into a text file with four columns
  \item{cluster }{ the name of the transcription cluster }
  \item{maxLRS }{ maximum of likelihood ratio statistic across all the covariates in xData }
  \item{n }{ total number of exon sets across all samples }
  \item{p }{ the number of candidate isoforms }
  \item{k0 }{ number of selected isoforms under null }
  \item{k1 }{ number of selected isoforms under alternative }
  \item{k0p }{ median number of selected isoforms under null across permutations or bootstraps}
  \item{k1p }{ median number of selected isoforms under alternative across permutations or bootstraps }
  \item{pResample}{resampling-based p-value}
  \item{nResample}{number of resamplings }
}
\references{
  Sun et al. IsoDOT Detects Differential RNA-isoform expression with respect to a 
  Categorical or Continuous Covariate with High Sensitivity and Specificity.
}
\author{ Wei Sun weisun@email.unc.edu }
\examples{
	
data(isoDuEx2)

names(isoDuEx2$gL[[1]])

g2test     = "chr19_490"
tags       = names(isoDuEx2$gL)
xData      = isoDuEx2$xData[,1:2]
readDepth  = isoDuEx2$readDepth$rd
pdDistL    = isoDuEx2$pdDistL
outputFile = "eQTL_chr19_490_known_iso_duOnly.txt"

all(isoDuEx2$readDepth$sampleID == tags)

isoDu(tags, isoDuEx2$gL, xData, readDepth, outputFile, pdDistL,
g2test, readLen=37, method="permutation", lmax=300,
nResample=10, maxTime=1000, duOnly=TRUE)

}

\keyword{ methods }




