\name{isoform-package}
\alias{isoform-package}
\alias{isoform}
\docType{package}
\title{
  RNA isoform related methods
}
\description{
  RNA isoform detection, expression abundance estimation, and differenial 
  expression studies
}
\details{
\tabular{ll}{
Package: \tab isoform\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-26\cr
License: \tab GPL version 2 or later\cr
LazyLoad: \tab yes\cr
}
}
\author{ Wei Sun wsun@bios.unc.edu }
\references{
  Sun et al. A statistical method for simultaneous isoform detection 
  and assessment of differential alternative expression
}
\keyword{ package }
